+++ boss/Xoryn.lua
@@ -6,9 +6,16 @@ LCH.Xoryn = {
     knot_will_drop_at = 0,
     knot_counter = 0,
     knot_dropped = false,
+    knot_picked_at = 0,
+    knot_picked_by = nil,
     arcane_conveyance_casted_at = 0,
     arcane_conveyance_on_player = false,
     arcane_conveyance_on_player_at = 0,
+
+    weakening_charge_on_player = false,
+    weakening_charge_endtime = 0,
+
+    last_accelerating_charge_at = 0,
     -- boss_is_light_colored = true,
     -- mirror_drain_endtime = 0,
     -- mirror_drain_stacks = 0,
@@ -39,23 +46,36 @@ function LCH.Xoryn.Reset()
     data.knot_will_drop_at = 0
     data.knot_counter = 0
     data.knot_dropped = false
+    data.knot_picked_at = 0
+    data.knot_picked_by = nil
   end
 
   data.arcane_conveyance_casted_at = 0
   data.arcane_conveyance_on_player_at = 0
+  data.weakening_charge_on_player = false
+  data.weakening_charge_endtime = 0
+  data.last_accelerating_charge_at = 0
 
 end
 
 
-function LCH.Xoryn.OnKnotPick(endTime)
+function LCH.Xoryn.OnKnotPick(endTime, unitName)
+  local now = GetGameTimeSeconds()
+  if GetGameTimeSeconds() - data.knot_picked_at < 10 and data.knot_picked_by == unitName then
+    -- filter out double event on combat start with first knot
+    return
+  end
+
   data.knot_active = true
   data.knot_will_drop_at = endTime
   data.knot_counter = data.knot_counter + 1
   data.knot_dropped = false
+  data.knot_picked_at = now
+  data.knot_picked_by = unitName
 end
 
 
-function LCH.Xoryn.OnKnotDrop()
+function LCH.Xoryn.OnKnotDrop(unitName)
   data.knot_will_drop_at = 0
   data.knot_dropped = true
 end
@@ -69,6 +89,19 @@ function LCH.Xoryn.ArcaneConveyanceDebuff(unitTag, isGain)
     data.arcane_conveyance_on_player = isGain
     data.arcane_conveyance_on_player_at = GetGameTimeSeconds()
   end
+end
+
+function LCH.Xoryn.WeakeningChargeDebuff(isGain, endTime)
+  data.weakening_charge_on_player = isGain
+  data.weakening_charge_endtime = endTime
+end
+
+
+function LCH.Xoryn.XorynAcceleratingCharge()
+  if LCH.savedVariables.showAcceleratingCharge then
+    CombatAlerts.Alert("", "Lightning", 0x345eebff, nil, 1500)
+  end
+  data.last_accelerating_charge_at = GetGameTimeSeconds()
 end
 
 
@@ -112,6 +145,21 @@ function LCH.Xoryn.UpdateTick(timeSec)
     LCHMessage2Label:SetHidden(true)
   end
 
+  if data.weakening_charge_on_player then
+    local debuffTimer = data.weakening_charge_endtime - timeSec
+    if debuffTimer > 0 then
+      LCHStatusLabelWeakeningChargeValue:SetText(string.format("%.0f s", debuffTimer))
+    else
+      LCHStatusLabelWeakeningChargeValue:SetText("-")
+    end
+  end
+
+  local nextChainLightning = 16 - (GetGameTimeSeconds() - data.last_accelerating_charge_at)
+  if nextChainLightning > 0 then
+    LCHStatusLabelAcceleratingChargeValue:SetText(string.format("%.0fs", nextChainLightning))
+  else
+    LCHStatusLabelAcceleratingChargeValue:SetText("-")
+  end
 
   -- Display UI elements needed for this boss.
   LCHStatus:SetHidden(false)
@@ -119,6 +167,10 @@ function LCH.Xoryn.UpdateTick(timeSec)
   LCHStatusLabelKnotTimerValue:SetHidden(not LCH.savedVariables.showKnotDropTimer)
   LCHStatusLabelKnotCounter:SetHidden(not LCH.savedVariables.showKnotCounter)
   LCHStatusLabelKnotCounterValue:SetHidden(not LCH.savedVariables.showKnotCounter)
+  LCHStatusLabelWeakeningCharge:SetHidden(not (LCH.savedVariables.showWeakeningCharge and LCH.status.isXoryn))
+  LCHStatusLabelWeakeningChargeValue:SetHidden(not (LCH.savedVariables.showWeakeningCharge and LCH.status.isXoryn))
+  LCHStatusLabelAcceleratingCharge:SetHidden(not (LCH.savedVariables.showAcceleratingCharge and LCH.status.isXoryn))
+  LCHStatusLabelAcceleratingChargeValue:SetHidden(not (LCH.savedVariables.showAcceleratingCharge and LCH.status.isXoryn))
 
 end